requirements.txt
fastapi
uvicorn[standard]
docker
python-jose[cryptography]
python-multipart
passlib[bcrypt]
python-dotenv